Refactor board member component into its own file.
