+++ app/styles/App.scss
@@ -2,6 +2,7 @@
 @import 'constants';
 @import 'AddComment';
 @import 'AddQuote';
+@import 'BoardMembersEditor';
 @import 'BoardView';
 @import 'Comment';
 @import 'LikeIcon';
